make strategies be part of configs, not assays. Delete custom assays, keep strategies
