+++ core/config/basicFungibleConfig.js
@@ -2,7 +2,7 @@ import harden from '@agoric/harden';
 
 import { noCustomization } from './noCustomization';
 import { makeCoreMintKeeper } from './coreMintKeeper';
-import { makeNatAssay } from './natAssay';
+import natStrategy from './strategies/natStrategy';
 
 // Fungible tokens (our default for mints) do not customize
 // payments, purses, etc. They use the "basic" mintKeeper (the place
@@ -16,7 +16,7 @@ function makeBasicFungibleConfig() {
   return harden({
     ...noCustomization,
     makeMintKeeper: makeCoreMintKeeper,
-    makeAssay: makeNatAssay,
+    strategy: natStrategy,
   });
 }
