+++ core/config/strategies/natStrategy.js
@@ -1,7 +1,5 @@
+import Nat from '@agoric/nat';
 import harden from '@agoric/harden';
-
-import natStrategy from './natStrategy';
-import { makeAssayMaker } from '../assay';
 
 // The default kind of amount is a labeled natural number describing a
 // quantity of fungible erights. The label describes what kinds of
@@ -12,8 +10,15 @@ import { makeAssayMaker } from '../assay';
 // rounding issues make floats problematic. All operations should be
 // done with the smallest whole unit such that the NatAssay never
 // deals with fractional parts.
-const makeNatAssay = makeAssayMaker(natStrategy);
 
-harden(makeNatAssay);
+const natStrategy = harden({
+  insistKind: Nat,
+  empty: _ => 0,
+  isEmpty: nat => nat === 0,
+  includes: (whole, part) => whole >= part,
+  equals: (left, right) => left === right,
+  with: (left, right) => Nat(left + right),
+  without: (whole, part) => Nat(whole - part),
+});
 
-export { makeNatAssay };
+export default natStrategy;
