+++ core/config/uniAssayConfig.js
@@ -2,7 +2,7 @@ import harden from '@agoric/harden';
 
 import { noCustomization } from './noCustomization';
 import { makeCoreMintKeeper } from './coreMintKeeper';
-import { makeUniAssayMaker } from './uniAssay';
+import { makeUniStrategy } from './strategies/uniStrategy';
 
 // This UniAssay config is used to create invites and similar assets.
 // It does not customize the purses, payments, mints, or issuers, and
@@ -12,7 +12,7 @@ function makeUniAssayConfigMaker(descriptionCoercer) {
     return harden({
       ...noCustomization,
       makeMintKeeper: makeCoreMintKeeper,
-      makeAssay: makeUniAssayMaker(descriptionCoercer),
+      strategy: makeUniStrategy(descriptionCoercer),
     });
   }
   return makeUniAssayConfig;
