+++ more/peg/peg.js
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 import harden from '@agoric/harden';
 import { makeNatAssay } from '../../core/config/natAssay';
 import { makeMint } from '../../core/issuers';
