+++ more/pixels/pixelStrategy.js
@@ -7,6 +7,10 @@ import {
   withoutPixelList,
 } from './types/pixelList';
 
+// A pixelList is a naive collection of pixels in the form:
+// [ { x: 0, y: 0 }, { x: 1, y: 1} ...]
+// This is less than ideal for efficiency and expressiveness but will
+// do for now
 const makePixelStrategy = canvasSize => {
   const pixelStrategy = harden({
     insistKind: pixelList => {
